Don't 'prepare' in GMO's purchase method
Ran into an issue where if an order which was previously submitted
to gmo ( duplicate order id ) we couldn't get past the prepare step.
one solution was to search for the order id and then go straight to
exectrans ('pay') but in order to avoid roundtrips we opted to save
the accessid and accesspass associated with the orderid in spree.

That code lives in our application and passes the accessid/accesspass
into the options hash parameter in the purchase method. this breaks
this gateway for anyone who wants to use it in their app but there
is a comment in gmo.rb which refers to the last working ref which
doesn't depend on our code.