Set money format to :cents
GMO will round down if it is sent a decimal value. So we now pass
authorize and prepare the final amounts in yen instead of the
amount*100.

The reason we used dollars in the first place is because this
branch is used in conjunction with spree and spree ha(s/d?) a bug
where it would send gateways the order amount * 100... and using
:dollars divides by 100..

...